Small/important change to track submit
The support documentation stated that, for us to use an input field as an identity, it had to match "userid," "login," "username," or "email." However, the input name simply has to contain one of those strings to be used as an identity. I tested it with both "signUpEmail" and "signUpEmailTest" to confirm that this was the case. 

A client ended up with identity splitting when we automatically identified users from a form field named "signUpEmail." We ended up giving them a refund because (as they pointed out) this documentation was incorrect.